Seeding data and migrations

+++ estate-agent-website/database/migrations/2024_03_31_221822_add_fields_to_listings_table.php
@@ -13,7 +13,7 @@ return new class extends Migration {
         Schema::table('listings', function (Blueprint $table) {
             $table->unsignedTinyInteger("beds");
             $table->unsignedTinyInteger("baths");
-            $table->unsignedTinyInteger("area");
+            $table->unsignedSmallInteger("area");
 
             $table->tinyText("city");
             $table->tinyText("code");
@@ -41,3 +41,4 @@ return new class extends Migration {
         ]);
     }
 };
+
